update types and mappers
