--- pkg/controller/direct/dataplex/mapper.generated.go
@@ -1,439 +1,0 @@
-// Copyright 2025 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// +generated:mapper
-// krm.group: dataplex.cnrm.cloud.google.com
-// krm.version: v1alpha1
-// proto.service: google.cloud.dataplex.v1
-
-package dataplex
-
-import (
-	pb "cloud.google.com/go/dataplex/apiv1/dataplexpb"
-	krmv1alpha1 "github.com/GoogleCloudPlatform/k8s-config-connector/apis/dataplex/v1alpha1"
-	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct"
-)
-
-func Content_FromProto(mapCtx *direct.MapContext, in *pb.Content) *krmv1alpha1.Content {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.Content{}
-	// MISSING: Name
-	// MISSING: Uid
-	out.Path = direct.LazyPtr(in.GetPath())
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Labels = in.Labels
-	out.Description = direct.LazyPtr(in.GetDescription())
-	out.DataText = direct.LazyPtr(in.GetDataText())
-	out.SQLScript = Content_SQLScript_FromProto(mapCtx, in.GetSqlScript())
-	out.Notebook = Content_Notebook_FromProto(mapCtx, in.GetNotebook())
-	return out
-}
-func Content_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.Content) *pb.Content {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Content{}
-	// MISSING: Name
-	// MISSING: Uid
-	out.Path = direct.ValueOf(in.Path)
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Labels = in.Labels
-	out.Description = direct.ValueOf(in.Description)
-	if oneof := Content_DataText_ToProto(mapCtx, in.DataText); oneof != nil {
-		out.Data = oneof
-	}
-	if oneof := Content_SQLScript_ToProto(mapCtx, in.SQLScript); oneof != nil {
-		out.Content = &pb.Content_SqlScript_{SqlScript: oneof}
-	}
-	if oneof := Content_Notebook_ToProto(mapCtx, in.Notebook); oneof != nil {
-		out.Content = &pb.Content_Notebook_{Notebook: oneof}
-	}
-	return out
-}
-func ContentObservedState_FromProto(mapCtx *direct.MapContext, in *pb.Content) *krmv1alpha1.ContentObservedState {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.ContentObservedState{}
-	out.Name = direct.LazyPtr(in.GetName())
-	out.Uid = direct.LazyPtr(in.GetUid())
-	// MISSING: Path
-	out.CreateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetCreateTime())
-	out.UpdateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetUpdateTime())
-	// MISSING: Labels
-	// MISSING: Description
-	// MISSING: DataText
-	// MISSING: SQLScript
-	// MISSING: Notebook
-	return out
-}
-func ContentObservedState_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.ContentObservedState) *pb.Content {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Content{}
-	out.Name = direct.ValueOf(in.Name)
-	out.Uid = direct.ValueOf(in.Uid)
-	// MISSING: Path
-	out.CreateTime = direct.StringTimestamp_ToProto(mapCtx, in.CreateTime)
-	out.UpdateTime = direct.StringTimestamp_ToProto(mapCtx, in.UpdateTime)
-	// MISSING: Labels
-	// MISSING: Description
-	// MISSING: DataText
-	// MISSING: SQLScript
-	// MISSING: Notebook
-	return out
-}
-func DataplexEntryTypeSpec_FromProto(mapCtx *direct.MapContext, in *pb.EntryType) *krmv1alpha1.DataplexEntryTypeSpec {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.DataplexEntryTypeSpec{}
-	// MISSING: Name
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	// MISSING: Description
-	// MISSING: DisplayName
-	// MISSING: Labels
-	// MISSING: Etag
-	// MISSING: TypeAliases
-	// MISSING: Platform
-	// MISSING: System
-	out.RequiredAspects = direct.Slice_FromProto(mapCtx, in.RequiredAspects, EntryType_AspectInfo_FromProto)
-	out.Authorization = EntryType_Authorization_FromProto(mapCtx, in.GetAuthorization())
-	return out
-}
-func DataplexEntryTypeSpec_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.DataplexEntryTypeSpec) *pb.EntryType {
-	if in == nil {
-		return nil
-	}
-	out := &pb.EntryType{}
-	// MISSING: Name
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	// MISSING: Description
-	// MISSING: DisplayName
-	// MISSING: Labels
-	// MISSING: Etag
-	// MISSING: TypeAliases
-	// MISSING: Platform
-	// MISSING: System
-	out.RequiredAspects = direct.Slice_ToProto(mapCtx, in.RequiredAspects, EntryType_AspectInfo_ToProto)
-	out.Authorization = EntryType_Authorization_ToProto(mapCtx, in.Authorization)
-	return out
-}
-func EntryType_FromProto(mapCtx *direct.MapContext, in *pb.EntryType) *krmv1alpha1.EntryType {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.EntryType{}
-	// MISSING: Name
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Description = direct.LazyPtr(in.GetDescription())
-	out.DisplayName = direct.LazyPtr(in.GetDisplayName())
-	out.Labels = in.Labels
-	out.Etag = direct.LazyPtr(in.GetEtag())
-	out.TypeAliases = in.TypeAliases
-	out.Platform = direct.LazyPtr(in.GetPlatform())
-	out.System = direct.LazyPtr(in.GetSystem())
-	out.RequiredAspects = direct.Slice_FromProto(mapCtx, in.RequiredAspects, EntryType_AspectInfo_FromProto)
-	out.Authorization = EntryType_Authorization_FromProto(mapCtx, in.GetAuthorization())
-	return out
-}
-func EntryType_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.EntryType) *pb.EntryType {
-	if in == nil {
-		return nil
-	}
-	out := &pb.EntryType{}
-	// MISSING: Name
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Description = direct.ValueOf(in.Description)
-	out.DisplayName = direct.ValueOf(in.DisplayName)
-	out.Labels = in.Labels
-	out.Etag = direct.ValueOf(in.Etag)
-	out.TypeAliases = in.TypeAliases
-	out.Platform = direct.ValueOf(in.Platform)
-	out.System = direct.ValueOf(in.System)
-	out.RequiredAspects = direct.Slice_ToProto(mapCtx, in.RequiredAspects, EntryType_AspectInfo_ToProto)
-	out.Authorization = EntryType_Authorization_ToProto(mapCtx, in.Authorization)
-	return out
-}
-func EntryTypeObservedState_FromProto(mapCtx *direct.MapContext, in *pb.EntryType) *krmv1alpha1.EntryTypeObservedState {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.EntryTypeObservedState{}
-	out.Name = direct.LazyPtr(in.GetName())
-	out.Uid = direct.LazyPtr(in.GetUid())
-	out.CreateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetCreateTime())
-	out.UpdateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetUpdateTime())
-	// MISSING: Description
-	// MISSING: DisplayName
-	// MISSING: Labels
-	// MISSING: Etag
-	// MISSING: TypeAliases
-	// MISSING: Platform
-	// MISSING: System
-	// MISSING: RequiredAspects
-	// MISSING: Authorization
-	return out
-}
-func EntryTypeObservedState_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.EntryTypeObservedState) *pb.EntryType {
-	if in == nil {
-		return nil
-	}
-	out := &pb.EntryType{}
-	out.Name = direct.ValueOf(in.Name)
-	out.Uid = direct.ValueOf(in.Uid)
-	out.CreateTime = direct.StringTimestamp_ToProto(mapCtx, in.CreateTime)
-	out.UpdateTime = direct.StringTimestamp_ToProto(mapCtx, in.UpdateTime)
-	// MISSING: Description
-	// MISSING: DisplayName
-	// MISSING: Labels
-	// MISSING: Etag
-	// MISSING: TypeAliases
-	// MISSING: Platform
-	// MISSING: System
-	// MISSING: RequiredAspects
-	// MISSING: Authorization
-	return out
-}
-func EntryType_AspectInfo_FromProto(mapCtx *direct.MapContext, in *pb.EntryType_AspectInfo) *krmv1alpha1.EntryType_AspectInfo {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.EntryType_AspectInfo{}
-	out.Type = direct.LazyPtr(in.GetType())
-	return out
-}
-func EntryType_AspectInfo_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.EntryType_AspectInfo) *pb.EntryType_AspectInfo {
-	if in == nil {
-		return nil
-	}
-	out := &pb.EntryType_AspectInfo{}
-	out.Type = direct.ValueOf(in.Type)
-	return out
-}
-func EntryType_Authorization_FromProto(mapCtx *direct.MapContext, in *pb.EntryType_Authorization) *krmv1alpha1.EntryType_Authorization {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.EntryType_Authorization{}
-	out.AlternateUsePermission = direct.LazyPtr(in.GetAlternateUsePermission())
-	return out
-}
-func EntryType_Authorization_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.EntryType_Authorization) *pb.EntryType_Authorization {
-	if in == nil {
-		return nil
-	}
-	out := &pb.EntryType_Authorization{}
-	out.AlternateUsePermission = direct.ValueOf(in.AlternateUsePermission)
-	return out
-}
-func Environment_FromProto(mapCtx *direct.MapContext, in *pb.Environment) *krmv1alpha1.Environment {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.Environment{}
-	// MISSING: Name
-	out.DisplayName = direct.LazyPtr(in.GetDisplayName())
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Labels = in.Labels
-	out.Description = direct.LazyPtr(in.GetDescription())
-	// MISSING: State
-	out.InfrastructureSpec = Environment_InfrastructureSpec_FromProto(mapCtx, in.GetInfrastructureSpec())
-	out.SessionSpec = Environment_SessionSpec_FromProto(mapCtx, in.GetSessionSpec())
-	// MISSING: SessionStatus
-	// MISSING: Endpoints
-	return out
-}
-func Environment_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.Environment) *pb.Environment {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Environment{}
-	// MISSING: Name
-	out.DisplayName = direct.ValueOf(in.DisplayName)
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Labels = in.Labels
-	out.Description = direct.ValueOf(in.Description)
-	// MISSING: State
-	out.InfrastructureSpec = Environment_InfrastructureSpec_ToProto(mapCtx, in.InfrastructureSpec)
-	out.SessionSpec = Environment_SessionSpec_ToProto(mapCtx, in.SessionSpec)
-	// MISSING: SessionStatus
-	// MISSING: Endpoints
-	return out
-}
-func EnvironmentObservedState_FromProto(mapCtx *direct.MapContext, in *pb.Environment) *krmv1alpha1.EnvironmentObservedState {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.EnvironmentObservedState{}
-	out.Name = direct.LazyPtr(in.GetName())
-	// MISSING: DisplayName
-	out.Uid = direct.LazyPtr(in.GetUid())
-	out.CreateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetCreateTime())
-	out.UpdateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetUpdateTime())
-	// MISSING: Labels
-	// MISSING: Description
-	out.State = direct.Enum_FromProto(mapCtx, in.GetState())
-	// MISSING: InfrastructureSpec
-	// MISSING: SessionSpec
-	out.SessionStatus = Environment_SessionStatusObservedState_FromProto(mapCtx, in.GetSessionStatus())
-	out.Endpoints = Environment_EndpointsObservedState_FromProto(mapCtx, in.GetEndpoints())
-	return out
-}
-func EnvironmentObservedState_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.EnvironmentObservedState) *pb.Environment {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Environment{}
-	out.Name = direct.ValueOf(in.Name)
-	// MISSING: DisplayName
-	out.Uid = direct.ValueOf(in.Uid)
-	out.CreateTime = direct.StringTimestamp_ToProto(mapCtx, in.CreateTime)
-	out.UpdateTime = direct.StringTimestamp_ToProto(mapCtx, in.UpdateTime)
-	// MISSING: Labels
-	// MISSING: Description
-	out.State = direct.Enum_ToProto[pb.State](mapCtx, in.State)
-	// MISSING: InfrastructureSpec
-	// MISSING: SessionSpec
-	out.SessionStatus = Environment_SessionStatusObservedState_ToProto(mapCtx, in.SessionStatus)
-	out.Endpoints = Environment_EndpointsObservedState_ToProto(mapCtx, in.Endpoints)
-	return out
-}
-func Environment_Endpoints_FromProto(mapCtx *direct.MapContext, in *pb.Environment_Endpoints) *krmv1alpha1.Environment_Endpoints {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.Environment_Endpoints{}
-	// MISSING: Notebooks
-	// MISSING: SQL
-	return out
-}
-func Environment_Endpoints_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.Environment_Endpoints) *pb.Environment_Endpoints {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Environment_Endpoints{}
-	// MISSING: Notebooks
-	// MISSING: SQL
-	return out
-}
-func Environment_SessionStatus_FromProto(mapCtx *direct.MapContext, in *pb.Environment_SessionStatus) *krmv1alpha1.Environment_SessionStatus {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.Environment_SessionStatus{}
-	// MISSING: Active
-	return out
-}
-func Environment_SessionStatus_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.Environment_SessionStatus) *pb.Environment_SessionStatus {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Environment_SessionStatus{}
-	// MISSING: Active
-	return out
-}
-func Lake_FromProto(mapCtx *direct.MapContext, in *pb.Lake) *krmv1alpha1.Lake {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.Lake{}
-	// MISSING: Name
-	out.DisplayName = direct.LazyPtr(in.GetDisplayName())
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Labels = in.Labels
-	out.Description = direct.LazyPtr(in.GetDescription())
-	// MISSING: State
-	// MISSING: ServiceAccount
-	out.Metastore = Lake_Metastore_FromProto(mapCtx, in.GetMetastore())
-	// MISSING: AssetStatus
-	// MISSING: MetastoreStatus
-	return out
-}
-func Lake_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.Lake) *pb.Lake {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Lake{}
-	// MISSING: Name
-	out.DisplayName = direct.ValueOf(in.DisplayName)
-	// MISSING: Uid
-	// MISSING: CreateTime
-	// MISSING: UpdateTime
-	out.Labels = in.Labels
-	out.Description = direct.ValueOf(in.Description)
-	// MISSING: State
-	// MISSING: ServiceAccount
-	out.Metastore = Lake_Metastore_ToProto(mapCtx, in.Metastore)
-	// MISSING: AssetStatus
-	// MISSING: MetastoreStatus
-	return out
-}
-func LakeObservedState_FromProto(mapCtx *direct.MapContext, in *pb.Lake) *krmv1alpha1.LakeObservedState {
-	if in == nil {
-		return nil
-	}
-	out := &krmv1alpha1.LakeObservedState{}
-	out.Name = direct.LazyPtr(in.GetName())
-	// MISSING: DisplayName
-	out.Uid = direct.LazyPtr(in.GetUid())
-	out.CreateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetCreateTime())
-	out.UpdateTime = direct.StringTimestamp_FromProto(mapCtx, in.GetUpdateTime())
-	// MISSING: Labels
-	// MISSING: Description
-	out.State = direct.Enum_FromProto(mapCtx, in.GetState())
-	out.ServiceAccount = direct.LazyPtr(in.GetServiceAccount())
-	// MISSING: Metastore
-	out.AssetStatus = AssetStatus_FromProto(mapCtx, in.GetAssetStatus())
-	out.MetastoreStatus = Lake_MetastoreStatus_FromProto(mapCtx, in.GetMetastoreStatus())
-	return out
-}
-func LakeObservedState_ToProto(mapCtx *direct.MapContext, in *krmv1alpha1.LakeObservedState) *pb.Lake {
-	if in == nil {
-		return nil
-	}
-	out := &pb.Lake{}
-	out.Name = direct.ValueOf(in.Name)
-	// MISSING: DisplayName
-	out.Uid = direct.ValueOf(in.Uid)
-	out.CreateTime = direct.StringTimestamp_ToProto(mapCtx, in.CreateTime)
-	out.UpdateTime = direct.StringTimestamp_ToProto(mapCtx, in.UpdateTime)
-	// MISSING: Labels
-	// MISSING: Description
-	out.State = direct.Enum_ToProto[pb.State](mapCtx, in.State)
-	out.ServiceAccount = direct.ValueOf(in.ServiceAccount)
-	// MISSING: Metastore
-	out.AssetStatus = AssetStatus_ToProto(mapCtx, in.AssetStatus)
-	out.MetastoreStatus = Lake_MetastoreStatus_ToProto(mapCtx, in.MetastoreStatus)
-	return out
-}
